pyproject.toml
[project]
name = "mcp-writing-library"
version = "1.0.0"
description = "MCP server for writing passages and terminology dictionary with hybrid semantic search"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Danilo da Silva" }
]

dependencies = [
    "kbase-core>=1.3.0",
    "mcp>=1.13.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv.sources]
kbase-core = { path = "../libraries/kbase-core", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
